+++ src/http/request/request.cpp
@@ -1,4 +1,5 @@
 #include "request.hpp"
+#include "io_pending_state.hpp"
 
 namespace http {
 void Request::run() {
